// packages/tx-categorize/src/enums.ts
export enum Action {
  APPROVE = 'APPROVE',
  BORROW = 'BORROW',
  UNWRAP = 'UNWRAP',
  DOMAIN_RENEW = 'DOMAIN_RENEW',
  DEPOSIT_TOKEN = 'DEPOSIT_TOKEN',
  REPAY = 'REPAY',
  MINT = 'MINT',
  EXCHANGE = 'EXCHANGE',
  LENDING = 'LENDING',
  BRIDGE_IN = 'BRIDGE_IN',
  BRIDGE_OUT = 'BRIDGE_OUT',
  CLAIM = 'CLAIM',
  DOMAIN_TRANSFER = 'DOMAIN_TRANSFER',
  DOMAIN_REGISTER = 'DOMAIN_REGISTER',
  DOMAIN_RESOLVER_SET = 'DOMAIN_RESOLVER_SET',
  VOTE = 'VOTE',
  STAKE = 'STAKE',
  REGISTER = 'REGISTER',
  CANCEL_ORDER = 'CANCEL_ORDER',
  BRIDGE = 'BRIDGE',
  MINING_PAYOUT = 'MINING_PAYOUT',
  DEPOSIT = 'DEPOSIT',
  DEPOSIT_MULTI = 'DEPOSIT_MULTI',
  WITHDRAW = 'WITHDRAW',
  WITHDRAW_MULTI = 'WITHDRAW_MULTI',
  TRANSFER = 'TRANSFER',
  EXCHANGE_FOR_MULTI = 'EXCHANGE_FOR_MULTI',
  EXCHANGE_FROM_MULTI = 'EXCHANGE_FROM_MULTI',
  NFT_EXCHANGE = 'NFT_EXCHANGE',
  NFT_APPROVE = 'NFT_APPROVE',
  NFT_TRANSFER = 'NFT_TRANSFER',
  MULTISIG_APPROVE_TX = 'MULTISIG_APPROVE_TX',
  DEPOSIT_ETH = 'DEPOSIT_ETH',
  WRAP = 'WRAP',
  WITHDRAW_TOKEN = 'WITHDRAW_TOKEN',
  RECLAIM = 'RECLAIM',
  PAYMENT = 'PAYMENT',
  CONTRACT_CALL = 'CONTRACT_CALL',
  STANDARD = 'STANDARD',
  STANDARD_SENT = 'STANDARD_SENT',
  STANDARD_RECEIVED = 'STANDARD_RECEIVED',
}